refactor(mobile): route web consumers through native helpers
Move duplicated builtin invocation, catalog command, composer input, rail state, and model catalog type logic into the pure helper modules introduced by the native prerequisite commits. Update existing consumers and focused tests to import the shared implementations while preserving public catalog command and model catalog exports.
