+++ cmd/evener-hub/frontend/src/stores/threads.ts
@@ -48,6 +48,8 @@ import {
 import { MutationOutboxIndexedDB } from "./mutationOutboxIndexedDB";
 import { createSecureUUID } from "./secureUUID";
 import { resetTasksPanelStoreForTests } from "./tasksPanel";
+import { buildInput, type InputAttachment } from "./composerInput";
+export type { InputAttachment } from "./composerInput";
 
 // InputAttachment is this store's real-attachment shape: base64 bytes, not a
 // hosted URL. The wire's InputItem (appwire/types.go:561-570) supports EITHER
@@ -66,13 +68,6 @@ import { resetTasksPanelStoreForTests } from "./tasksPanel";
 // durable outbox record, the recovery draft that rebuilds a composer - pairs
 // text and attachment by identity instead of re-deriving the pairing from
 // array position. buildInput drops it when it assembles the wire input.
-export interface InputAttachment {
-  marker: number;
-  mediaType: string;
-  data: string; // base64-encoded bytes (wire InputItem.data)
-  name?: string;
-}
-
 export type ComposerMutationRoute = "send" | "queue" | "steer" | "drain";
 
 // ForkFromTurnOptions mirrors ThreadForkParams verbatim (appwire/types.go:
@@ -1132,17 +1127,6 @@ export function appendFrameTime(times: number[], now: number): number[] {
 // rejects the call), then one image item per attachment. The text arrives
 // verbatim: any new SUBMIT path through here owes it the same
 // translateAttachmentMarkers pass composerMutationIntent applies.
-function buildInput(text: string, attachments?: InputAttachment[]): InputItem[] {
-  const input: InputItem[] = [];
-  if (text.trim()) input.push({ type: "text", text });
-  for (const att of attachments ?? []) {
-    const image: InputItem = { type: "image", mediaType: att.mediaType, data: att.data };
-    if (att.name !== undefined) image.name = att.name;
-    input.push(image);
-  }
-  return input;
-}
-
 function attachmentBlob(attachment: InputAttachment): Blob {
   const bytes = Uint8Array.from(atob(attachment.data), (character) => character.charCodeAt(0));
   return new Blob([bytes], { type: attachment.mediaType });
