+++ cmd/evener-hub/frontend/src/widgets/modelCatalog/index.tsx
@@ -12,7 +12,7 @@
 // lines, and a list expanded the moment it opens.
 import { type JSX, type KeyboardEvent, useEffect, useId, useMemo, useRef, useState } from "react";
 import { friendlyLaunchErrorMessage } from "../../protocol/errors";
-import type { ModelDescriptor, ModelListDiagnostic } from "../../protocol/types.gen";
+import type { ModelDescriptor } from "../../protocol/types.gen";
 // Import siblings directly, never through the widgets barrel: this module is
 // itself barrel-exported, so importing the barrel here would be a cycle (the
 // same reason collectioneditor imports ../button directly).
@@ -22,6 +22,8 @@ import { Popover } from "../popover";
 import { Skeleton } from "../skeleton";
 import styles from "./modelCatalog.module.css";
 import { buildPickerRows, pickableRows } from "./pickerRows";
+import type { ModelCatalog, ModelCatalogDiagnostic, ModelCatalogEntry } from "./types";
+export type { ModelCatalog, ModelCatalogDiagnostic, ModelCatalogEntry } from "./types";
 
 const CLASS = {
   trigger: requireClass(styles.trigger, "modelCatalog.module.css", "trigger"),
@@ -57,15 +59,6 @@ const SKELETON_LINES = 4;
 // makes it optional because daemon/source callers may know only an identity.
 // Keeping the generated type as the source of truth prevents this view model
 // from drifting when the model/list contract changes.
-export type ModelCatalogEntry = Omit<ModelDescriptor, "displayName"> & { displayName: string };
-export type ModelCatalogDiagnostic = ModelListDiagnostic;
-
-export interface ModelCatalog {
-  models: ModelCatalogEntry[];
-  recent: ModelCatalogEntry[];
-  diagnostics?: ModelCatalogDiagnostic[];
-}
-
 export interface ModelCatalogProps {
   value: string;
   onChange: (qualified: string) => void;
